Move Context C++ declaration to new header file
